Added common test data to conftest.py
